+++ src/Global.css
@@ -68,9 +68,9 @@ button {
 * {
   -ms-overflow-style: none;
 }
-::-webkit-scrollbar {
+/* ::-webkit-scrollbar {
   display: none;
-}
+} */
 
 html,
 body,
