+++ src/components/Column/Card.module.css
@@ -1,0 +1,7 @@
+.container {
+  width: 100%;
+  padding: 20px;
+  border-radius: 6px;
+  border: 1px solid var(--gray-gray_D9D9D9, #d9d9d9);
+  background: var(--white-white_FFFFFF, #fff);
+}
