+++ src/components/Column/Card.tsx
@@ -1,5 +1,23 @@
-import React from "react";
+import styles from "./Card.module.css";
+import { CardType } from "../../interface/DashboardType";
 
-export default function Card() {
-  return <div></div>;
+export default function Card({ card }: { card: CardType }) {
+  return (
+    <div className={styles.container}>
+      <div>{card.title}</div>
+      <div>{card.assignee.nickname}</div>
+      <div>{card.dueDate}</div>
+      <div>{card.tags}</div>
+      {/* <div>와</div>
+      <div>와</div>
+      <div>와</div>
+      <div>와</div>
+      <div>와</div>
+      <div>와</div>
+      <div>와</div>
+      <div>와</div>
+      <div>와</div>
+      <div>와</div> */}
+    </div>
+  );
 }
